Character 코드 수정 및 밸런스 수정

+++ textRPG/Source/Game/IJob.h
@@ -8,8 +8,8 @@ public:
 	virtual ~IJob() = default;
 
 	// 순수 가상함수
-	virtual float GetPowerWeight() const = 0;
-	virtual float GetDefenseWeight() const = 0;
-	virtual float GetLuckWeight() const = 0;
+	virtual int GetPowerWeight() const = 0;
+	virtual int GetDefenseWeight() const = 0;
+	virtual int GetLuckWeight() const = 0;
 	virtual std::wstring GetJobName() const = 0;
 };
